✨ feat : 파란색 초대하기 버튼 제작, scss파일명 대문자 변경, 프리티어 확장자 추가
