+++ .prettierrc.json
@@ -1,6 +1,6 @@
 {
   "singleQuote": true,
-  "jsxSingleQuote": true,
+  "jsxSingleQuote": false,
   "semi": true,
   "tabWidth": 2,
   "useTabs": false,
