+++ src/components/atoms/buttons/eventDashboardBtn/index.tsx
@@ -1,8 +1,7 @@
 import { ButtonProps } from '@/@types/type';
-import styles from './eventDashboardBtn.module.scss';
+import styles from './EventDashboardBtn.module.scss';
 import Image from 'next/image';
 import plus_png from '../../../../../public/assets/icon/plus.png';
-
 
 const EventDashboardBtn = ({ name, type, onClick }: ButtonProps) => {
   const classNames = (type: ButtonProps['type']) => {
@@ -24,7 +23,7 @@ const EventDashboardBtn = ({ name, type, onClick }: ButtonProps) => {
       {name}
       {type !== 'deleteDashboard' && (
         <div className={styles.icon}>
-          <Image src={plus_png} alt='plus' width={20} height={20} />
+          <Image src={plus_png} alt="plus" width={20} height={20} />
         </div>
       )}
     </button>
@@ -35,11 +34,10 @@ export default EventDashboardBtn;
 
 //사용법
 
-// <eventDashboardBtn name="새로운 컬럼 추가하기" type="addColumn"/>
-//
+// <EventDashboardBtn name="새로운 컬럼 추가하기" type="addColumn"/>
 
-// <eventDashboardBtn name="새로운 대시보드" type="newDashboard"/>
+// <EventDashboardBtn name="새로운 대시보드" type="newDashboard"/>
 
-// <eventDashboardBtn type="addTodo"/>
+// <EventDashboardBtn type="addTodo"/>
 
-// <eventDashboardBtn name="대시보드 삭제하기" type="deleteDashboard"/>
+// <EventDashboardBtn name="대시보드 삭제하기" type="deleteDashboard"/>
