+++ src/components/atoms/buttons/headerBtn/HeaderBtn.module.scss
@@ -9,7 +9,9 @@
   justify-content: center;
   align-items: center;
   gap: 5px;
-  img{
+  cursor: pointer;
+
+  img {
     display: none;
   }
 }
@@ -23,7 +25,7 @@
   @media (min-width: 769px) {
     width: 88px;
     height: 40px;
-    .editImg{
+    .editImg {
       display: block;
     }
   }
@@ -38,8 +40,31 @@
   @media (min-width: 769px) {
     width: 116px;
     height: 40px;
-    .addBoxImg{
+    .addBoxImg {
       display: block;
     }
+  }
+}
+
+.inviteBlue {
+  width: 86px;
+  height: 28px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 6px;
+  border-radius: 4px;
+  border: none;
+  background: var(--blue30, #026fb0);
+  color: var(--white, #ffffff);
+  font-size: 1.2rem;
+  font-weight: 500;
+  cursor: pointer;
+
+  @media (min-width: 769px) {
+    width: 105px;
+    height: 32px;
+    font-size: 1.4rem;
+    font-weight: 500;
   }
 }
