+++ src/components/atoms/buttons/pagenationBtn/index.tsx
@@ -1,5 +1,5 @@
 import { ButtonProps } from '@/@types/type';
-import styles from './pagenationBtn.module.scss';
+import styles from './PagenationBtn.module.scss';
 import forwardArrowIcon from '../../../../../public/assets/icon/forwardArrowIcon.svg';
 import leftArrowIcon from '../../../../../public/assets/icon/leftArrowIcon.svg';
 import Image from 'next/image';
@@ -8,12 +8,12 @@ const PagenationBtn = ({ onClick }: ButtonProps) => {
   return (
     <div>
       <button className={styles.pagenationBtn} onClick={onClick}>
-        <Image src={leftArrowIcon} alt='leftArrowIcon' width={16} height={16} />
+        <Image src={leftArrowIcon} alt="leftArrowIcon" width={16} height={16} />
       </button>
       <button className={styles.pagenationBtn} onClick={onClick}>
         <Image
           src={forwardArrowIcon}
-          alt='rightArrowIcon'
+          alt="rightArrowIcon"
           width={16}
           height={16}
         />
@@ -23,3 +23,5 @@ const PagenationBtn = ({ onClick }: ButtonProps) => {
 };
 
 export default PagenationBtn;
+
+//PagenationBtn 컴포넌트는 onClick만
